Refact files structure
